[builds] in CODEGEN, backport cudand changes from gg_tt.mad #679, include small formatting fixes
